# src/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Weynon_V2",
    "description": "Welcome to Weynon Api",
    "contact": {}
  },
  "host": "178.79.177.139:8848/api",
  "basePath": "/",
  "securityDefinitions": {
    "auth": {
      "type": "oauth2",
      "flow": "implicit",
      "authorizationUrl": "http://178.79.177.139:8848/api",
      "scopes": {}
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/auth/signup": {
      "post": {
        "description": "TODO: Add Description",
        "summary": "Signup",
        "tags": [
          "Auth"
        ],
        "operationId": "AuthSignupPost",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/SignupRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
    "/auth/currentUser": {
      "get": {
        "description": "TODO: Add Description",
        "summary": "Get Current User",
        "tags": [
          "Auth"
        ],
        "operationId": "AuthCurrentUserGet",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
    "/auth/send-phone-verification-code": {
      "post": {
        "description": "TODO: Add Description",
        "summary": "Send Phone Verification Code",
        "tags": [
          "Auth"
        ],
        "operationId": "AuthSendPhoneVerificationCodePost",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/SendPhoneVerificationCodeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
    "/auth/confirm-phone-verification-code": {
      "post": {
        "description": "TODO: Add Description",
        "summary": "Confirm Phone Verification Code",
        "tags": [
          "Auth"
        ],
        "operationId": "AuthConfirmPhoneVerificationCodePost",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ConfirmPhoneVerificationCodeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
    "/auth/resend-phone-verification-code": {
      "get": {
        "description": "TODO: Add Description",
        "summary": "Resend Phone Verification Code",
        "tags": [
          "Auth"
        ],
        "operationId": "AuthResendPhoneVerificationCodeGet",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
    "/auth/login-via-local": {
      "post": {
        "description": "TODO: Add Description",
        "summary": "Login Via Local",
        "tags": [
          "Auth"
        ],
        "operationId": "AuthLoginViaLocalPost",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/LoginViaLocalRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
    "/auth/refresh-token": {
      "get": {
        "description": "TODO: Add Description",
        "summary": "Refresh Token",
        "tags": [
          "Auth"
        ],
        "operationId": "AuthRefreshTokenGet",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
...
    "/auth/new-password-phone": {
      "post": {
        "description": "TODO: Add Description",
        "summary": "New Password reset via Phone",
        "tags": [
          "Auth"
        ],
        "operationId": "AuthNewPasswordPhonePost",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/NewPasswordresetviaPhoneRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
    "/auth/update-user-profile": {
      "patch": {
        "description": "TODO: Add Description",
        "summary": "Update User Profile",
        "tags": [
          "Auth"
        ],
        "operationId": "AuthUpdateUserProfilePatch",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "postNotification",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "peerNotification",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "privateMsgNotification",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "dob",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "gender",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
    "/auth/update-password": {
      "post": {
        "description": "TODO: Add Description",
        "summary": "Update Password",
        "tags": [
          "Auth"
        ],
        "operationId": "AuthUpdatePasswordPost",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdatePasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
    "SignupRequest": {
      "title": "SignupRequest",
      "example": {
        "email": "8axif.javed@gmail.com",
        "username": "8axif.javed",
        "gender": "male",
        "dob": "2018-12-12",
        "password": "8axif.javed@gmail.com",
        "confirmationPassword": "8axif.javed@gmail.com"
      },
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "dob": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "confirmationPassword": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "username",
        "gender",
        "dob",
        "password",
        "confirmationPassword"
      ]
    },
    "SendPhoneVerificationCodeRequest": {
      "title": "SendPhoneVerificationCodeRequest",
      "example": {
        "phonenumber": "+923170770556"
      },
      "type": "object",
      "properties": {
        "phonenumber": {
          "type": "string"
        }
      },
      "required": [
        "phonenumber"
      ]
    },
    "ConfirmPhoneVerificationCodeRequest": {
      "title": "ConfirmPhoneVerificationCodeRequest",
      "example": {
        "code": "4944"
      },
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        }
      },
      "required": [
        "code"
      ]
    },
    "LoginViaLocalRequest": {
      "title": "LoginViaLocalRequest",
      "example": {
        "identifier": "axif.javed@gmail.com",
        "password": "axif.javed@gmail.com"
      },
      "type": "object",
      "properties": {
        "identifier": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": [
        "identifier",
        "password"
      ]
    },
    "NewPasswordresetviaPhoneRequest": {
      "title": "NewPasswordresetviaPhoneRequest",
      "example": {
        "password": "axif.javed@gmail.com",
        "confirmationPassword": "axif.javed@gmail.com",
        "code": "3180"
      },
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "confirmationPassword": {
          "type": "string"
        },
        "code": {
          "type": "string"
        }
      },
      "required": [
        "password",
        "confirmationPassword",
        "code"
      ]
    },
    "UpdatePasswordRequest": {
      "title": "UpdatePasswordRequest",
      "example": {
        "oldPassword": "8axif.javed@gmail.com",
        "newPassword": "8axif.javed@gmail.com"
      },
      "type": "object",
      "properties": {
        "oldPassword": {
          "type": "string"
        },
        "newPassword": {
          "type": "string"
        }
      },
      "required": [
        "oldPassword",
        "newPassword"
      ]
    },
...
  },
...
}